refactor(chat): move isomorphic chat helpers out of server/ into lib/chat/
Three "use client" components and lib/export-chat.ts were importing runtime
values out of @/server: deriveStatusLine from server/chat-helpers, and
extractText / rowToUIMessage from server/chat/ui-messages.

Both modules are pure (their only imports are types plus lib/quiz and
lib/retrieval-tool-names), so nothing leaked today. But this is the exact
boundary this branch set out to establish: anything under server/ has to stay
off client import paths, or a server-only dependency added later gets pulled
into the browser bundle with nothing to catch it.

Moved both to lib/chat/ and updated all 13 import sites. No client file
imports a runtime value from @/server any more.
